Refactors setting zsh as default shell
Here's roughly how I might've written it:

1   if [[ -z "$ZSH_VERSION" ]] ; then
2     echo "Changing zsh to default shell"
3
4     if ! which zsh | grep -qxFf - /etc/shells ; then
5       echo "Adding zsh to list of allowed shells"
6       which zsh | sudo tee -a /etc/shells > /dev/null
7     fi
8
9     chsh -s $(which zsh)
10  fi

Line 1: Test ZSH_VERSION to see if it's empty. If it is, you're not in
`zsh`.

Line 4: Check to see that the version of zsh you want to set as the
default is allowed as a login shell.

Line 6: If not, add it as an allowed shell.